+++ cg/constants/constants.py
@@ -102,16 +102,6 @@ class CancerAnalysisType(StrEnum):
     TUMOR_WGS = auto()
 
 
-class LibraryPrepCategory(StrEnum):
-    COVID: str = "cov"
-    MICROBIAL: str = "mic"
-    READY_MADE_LIBRARY: str = "rml"
-    TARGETED_GENOME_SEQUENCING: str = "tgs"
-    WHOLE_EXOME_SEQUENCING: str = "wes"
-    WHOLE_GENOME_SEQUENCING: str = "wgs"
-    WHOLE_TRANSCRIPTOME_SEQUENCING: str = "wts"
-
-
 class SexOptions(StrEnum):
     MALE: str = "male"
     FEMALE: str = "female"
